+++ static/js/components/EmbedlyCard.js
@@ -3,8 +3,7 @@
 import React from "react"
 import isURL from "validator/lib/isURL"
 
-import { loadEmbedlyPlatform } from "../lib/embed"
-import ReactDOMServer from "react-dom/server"
+import { loadEmbedlyPlatform, renderEmbedlyCard } from "../lib/embed"
 
 type Props = {
   url: string
@@ -21,16 +20,6 @@ export default class EmbedlyCard extends React.Component<Props> {
       return null
     }
 
-    const html = ReactDOMServer.renderToStaticMarkup(
-      <a
-        data-card-chrome="0"
-        data-card-controls="0"
-        data-card-key={SETTINGS.embedlyKey}
-        href={url}
-        className="embedly-card"
-      />
-    )
-
-    return <div dangerouslySetInnerHTML={{ __html: html }} />
+    return <div dangerouslySetInnerHTML={{ __html: renderEmbedlyCard(url) }} />
   }
 }
